feat(worker) : add profile page in worker
