+++ package.json
@@ -19,6 +19,7 @@
     "react-hook-form": "^7.51.4",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.2.1",
+    "react-quilljs": "^2.0.0",
     "react-responsive-carousel": "^3.2.23",
     "react-router-dom": "^6.23.1"
   },
