+++ src/main.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import "@/styles/index.css";
+import "quill/dist/quill.snow.css";
 import "@fontsource/poppins";
 import "react-responsive-carousel/lib/styles/carousel.min.css";
 import { NextUIProvider } from "@nextui-org/react";
