+++ src/views/Worker/pages/WorkerProfilePage.jsx
@@ -1,8 +1,8 @@
 // eslint-disable-next-line no-unused-vars
 import React from "react";
 import BaseLayout from "@/layouts/BaseLayout";
-import ProfileInput from "../components/ProfileInput";
 import { Helmet } from "react-helmet-async";
+import ProfileInput from "@/components/common/ProfileInput";
 
 const WorkerProfilePage = () => {
   return (
@@ -12,7 +12,7 @@ const WorkerProfilePage = () => {
       </Helmet>
       <BaseLayout>
         <div className="min-h-screen px-10 py-3">
-          <ProfileInput />
+          <ProfileInput type="worker" />
         </div>
       </BaseLayout>
     </>
